gcp/requirements.txt
# config
python-dotenv==0.17.1

# pyAnVIL
pyAnVIL==0.0.8rc4

# GCP
Flask==1.1.4
gunicorn==20.1.0